requirements.txt
google
flask==1.1.1
gunicorn==19.9.0
absl-py==0.8.0
astor==0.8.0
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
gast==0.3.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
idna==2.8
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.2.1
protobuf==3.10.0
pyparsing==2.4.2
python-dateutil==2.8.0
PyYAML==5.1.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
slim==0.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
torch==1.3.1+cpu
--find-links https://download.pytorch.org/whl/torch_stable.html
torchvision==0.4.2+cpu
--find-links https://download.pytorch.org/whl/torch_stable.html
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
scikit-image==0.16.2
firebase_admin==3.2.0
urllib3==1.25.6
pyclustering
